requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fonttools==4.62.1
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
kagglehub==1.0.0
kagglesdk==0.1.16
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
